Add requirement of Digest::MD5 >= 2.50
Inline::Python does not use Digest::MD5, but Inline does.
Versions < 2.50 had a bug causing UTF-8 arguments to be downgraded. This happend to the Python code in t/06dict.t and is the reason why on my development machine, the test worked without the coding: utf-8, while for many others it didn't. This should fix the problem for good.