Filter symlinked duplicates of python executables in Makefile.PL
openSUSE adds a symlink to /usr/bin named /usr/bin/X11 which made us
find the python executable in two places. This caused Makefile.PL to
ask the user which python executable to use, when they were in fact
the same. We now filter the list by the real path to avoid this.