Fix memory corruption after passing a dict with a non-string key to Perl
When a Python dictionary contains a key that is not a string but an
Object, we use the equivalent of str(key) for generating the Perl hash
key. The code for this contained a superfluous Py_DECREF on this
generated string that could cause memory corruption and segfaults later
on.