Try to detect more object types.
> use Inline Python => <<'END';
>
> import numpy
>
> def fun(x):
>     range = numpy.arange(x)
>     return range
>
> END
>
> fun(10)

I've had a look and the reason seems to be that Py2Pl thinks the numbers in the
range are hashes, gets a list of keys and values to convert and then segfaults
when trying to free those lists. Apparently they are not reference counted.

The reason why Py2Pl thinks those numbers are hashes is that they really are
objects (numpy.int64) and objects in Python are really magic dictionaries,
just like in Perl. Ordinarily objects would be wrapped in
Inline::Python::Object proxies but Py2Pl doesn't recognize those numbers as
objects. That's because there's simply no real way to check if something is an
object. PyInstance_Check(obj) only returns true if it's an "old style instance"
from the Python 1 era. The Py_TPFLAGS_HEAPTYPE flag is set on most user-defined
objects, but fails on things like PyDate and apparently also some C extensions.
Ironically, the only information on how to properly check for objects I could
find was a discussion from 2004 where someone tried to add that to Inline::Python.

The segfault goes away when I reshuffle the checks in Py2Pl so we first check if
something looks like an Int. But then you'd get a plain old integer SV. That's
nice if you want do do all your further processing in Perl, but not so great if
you want to pass it back to some numpy function. And I don't know if there are
other types in numpy where the simple reshuffle won't work.

Apparently I can recognize those weird objects as PyMapping_Check returns true,
but PyMapping_Length returns -1 (failure). The latter does so when there's no
inst->ob_type->tp_as_mapping->mp_length function.