Prevent overriding Perl definitions on multiple use of Inline::Python.
When Inline::Python is used twice from the same Perl script, the
second invocation would try to redefine the functions already declared
on the first invocation and bind them to no code.  This makes those
functions not callable and generates warnings about functions at the
Perl level being redefined on the console.

This commit fixes both the redefinitions and the warnings.